fix(map): hex now update color and size on zoom in
